Refactored CSV Template download & upload.
