+++ api/src/functions/utils/RcdaHttpHeaders.ts
@@ -1,5 +1,6 @@
 
 enum RcdaHttpHeaders {
-    ContentType = "Content-Type"
+    ContentType = "Content-Type",
+    ContentDisposition = "Content-Disposition"
 }
 export default RcdaHttpHeaders;
